refactoring (#58)
* refactoring
