--- src/main/java/jp/igapyon/simpleodata4/oiyokan/h2/data/TinyH2Util.java
@@ -1,76 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- * 
- *   http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License. 
- */
-package jp.igapyon.simpleodata4.oiyokan.h2.data;
-
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-
-/**
- * h2 database 用の小さなユーティリティクラス
- */
-public class TinyH2Util {
-    private TinyH2Util() {
-    }
-
-    /**
-     * h2 データベースへのDB接続を取得します。
-     * 
-     * @return データベース接続。
-     */
-    public static Connection getH2Connection() {
-        Connection conn;
-        try {
-            Class.forName("org.h2.Driver");
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-            throw new IllegalArgumentException(e);
-        }
-        // SQL Server 互換モードで動作させる.
-        final var jdbcConnStr = "jdbc:h2:mem:localmetainfo;DB_CLOSE_DELAY=-1;DATABASE_TO_UPPER=FALSE;CASE_INSENSITIVE_IDENTIFIERS=TRUE;MODE=MSSQLServer";
-        // System.err.println("TRACE: DEMO: [connect jdbc] " + jdbcConnStr);
-        try {
-            conn = DriverManager.getConnection(//
-                    jdbcConnStr, "sa", "");
-        } catch (SQLException ex) {
-            ex.printStackTrace();
-            throw new IllegalArgumentException(ex);
-        }
-
-        return conn;
-    }
-
-    /**
-     * SQL検索プレースホルダの文字列を生成します。
-     * 
-     * @param count プレースホルダ数。
-     * @return プレースホルダ文字列。
-     */
-    public static String getQueryPlaceholderString(int count) {
-        String queryPlaceholder = "";
-        for (int col = 0; col < count; col++) {
-            if (col != 0) {
-                queryPlaceholder += ",";
-            }
-            queryPlaceholder += "?";
-        }
-
-        return queryPlaceholder;
-    }
-}
